* Added ability to set individual register expiration times and policies in ExpiringRegistry * Added getMethod and HashMethod in ReflectUtil, taking class parameters * ISQL now takes an int port instead of a short port

+++ SQL-Lib/src/ninja/egg82/sql/ISQL.java
@@ -10,7 +10,7 @@ import ninja.egg82.patterns.events.EventHandler;
 
 public interface ISQL {
 	//functions
-	void connect(String address, short port, String user, String pass, String dbName);
+	void connect(String address, int port, String user, String pass, String dbName);
 	void connect(String address, String user, String pass, String dbName);
 	void connect(String filePath);
